+++ changelog.md
@@ -2,3 +2,6 @@
 - [packmodules.cpp] `StateTransFcn`删除$D$矩阵.
 - [simulator.cpp] ADDED: `FLAG_REDUNDANT`.
 - [matmodules.cpp/hpp] ADDED: `MOutput`.
+
+# V2.1.2
+- [matmodules.cpp/hpp] BUGFIXED: `MOutput`.